#1141 Added: скрипт генерации и тестовые данные для заполнения БД,

+++ src/main/java/ru/parma/filesdistr/repos/FileRepository.java
@@ -3,7 +3,7 @@ package ru.parma.filesdistr.repos;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.parma.filesdistr.models.File;
 
-public interface FileRepository extends JpaRepository<File, Integer> {
+public interface FileRepository extends JpaRepository<File, Long> {
 
 
 }
